+++ src/wiki/urls.py
@@ -2,7 +2,13 @@ from django.utils.module_loading import import_string
 from wiki.compat import include, url
 from wiki.conf import settings
 from wiki.core.plugins import registry
+from wiki import sites
 from wiki.views import accounts, article, deleted_list
+
+
+urlpatterns = [
+    url(r'^', sites.site.urls),
+]
 
 
 class WikiURLPatterns: